rename src_code to src for best practice convention, fix pytest root source by creating .ini file
