+++ pytest.ini
@@ -1,0 +1,5 @@
+# Use this file for configuring pytest. The package manager file is in the
+# src directory. It is not recommended to use a pyproject.toml to set the tests
+# root source directory since doing so may not launch the pipfile correctly.
+[pytest]
+pythonpath = src
