+++ src/cluster/__init__.py
@@ -1,4 +1,4 @@
-from ..util import frame
+from util import frame
 
 from .base import Base
 from .lsf import Lsf
